+++ src/signalling/mod.rs
@@ -1,7 +1,7 @@
 //! [WebRTC] [signalling] related implementations.
 //!
-//! [WebRTC]: https://webrtcglossary.com/webrtc/
-//! [signalling]: https://webrtcglossary.com/signaling/
+//! [WebRTC]: https://webrtcglossary.com/webrtc
+//! [signalling]: https://webrtcglossary.com/signaling
 
 pub mod elements;
 pub mod participants;