+++ src/turn/coturn/repo.rs
@@ -10,10 +10,9 @@ use derive_more::{Display, From};
 use failure::Fail;
 use redis::{IntoConnectionInfo, RedisError};
 
-use crate::{
-    log::prelude as log,
-    turn::{IceUser, IceUsername},
-};
+use crate::log::prelude as log;
+
+use super::ice_user::{CoturnIceUser, IceUsername};
 
 /// Medea's [Coturn] realm name.
 const COTURN_REALM: &str = "medea";
@@ -63,7 +62,10 @@ impl TurnDatabase {
     ///
     /// Errors if unable to establish connection with database, or database
     /// request fails.
-    pub async fn insert(&self, user: &IceUser) -> Result<(), TurnDatabaseErr> {
+    pub async fn insert(
+        &self,
+        user: &CoturnIceUser,
+    ) -> Result<(), TurnDatabaseErr> {
         log::debug!("Store ICE user: {:?}", user);
 
         let key = user.user().redis_key();
@@ -110,7 +112,7 @@ impl IceUsername {
     }
 }
 
-impl IceUser {
+impl CoturnIceUser {
     /// Forms a [Coturn]'s [HMAC key] of this [`IceUser`].
     ///
     /// [HMAC key]: https://tinyurl.com/y33qa86c